src/requirements.txt
semantic-kernel[azure]
python-dotenv 
opentelemetry-sdk 
azure-monitor-opentelemetry-exporter